删除产品时的并发问题，同时增加删除成功/失败的提示信息

+++ web/admin/src/main/java/com/xhm/longxin/qth/web/admin/module/action/BuyProductAction.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.alibaba.citrus.service.form.CustomErrors;
 import com.alibaba.citrus.service.requestcontext.parser.ParameterParser;
+import com.alibaba.citrus.turbine.Context;
 import com.alibaba.citrus.turbine.Navigator;
 import com.alibaba.citrus.turbine.dataresolver.FormField;
 import com.alibaba.citrus.turbine.dataresolver.FormGroup;
@@ -30,8 +31,18 @@ public class BuyProductAction {
 	@Autowired
 	private BuyProductService buyProductService;
 
-	public void doDeleteProduct(@Param("id") Long id) {
-		buyProductService.deleteBuyProductById(id);
+	public void doDeleteProduct(@Param("id") Long id, Context context) {
+		try {
+			BuyProduct product = buyProductService.getBuyProductById(id);
+			if (product != null) {
+				buyProductService.deleteBuyProductById(id);
+				context.put("productDeleted", true);
+				return;
+			}
+		} catch (Exception e) {
+			context.put("productDeleted", false);
+		}
+		context.put("productDeleted", false);
 	}
 
 	public void doAuditProduct(
@@ -45,10 +56,11 @@ public class BuyProductAction {
 			err.setMessage("auditFail");
 			return;
 		}
-		//如果产品状态不是待审核，则返回
-		BuyProduct product=buyProductService.getBuyProductById(auditVO.getAuditId());
-		if(product!=null){
-			if(!ProductStatus.NEW.equals(product.getStatus())){
+		// 如果产品状态不是待审核，则返回
+		BuyProduct product = buyProductService.getBuyProductById(auditVO
+				.getAuditId());
+		if (product != null) {
+			if (!ProductStatus.NEW.equals(product.getStatus())) {
 				err.setMessage("auditStatusFail");
 				return;
 			}
